Short class name with interface now does not make a collision

+++ src/classNode.h
@@ -245,7 +245,7 @@ private:
             std::vector<int> *inheritedIndexes = new std::vector<int>(0)) const;
     std::vector<QString> getAttributePrintable() const;
     QString getMethodParametersPrintable(std::vector<MethodParameter> parameters) const;
-    void getMaxWidth(std::vector<QString> toCompare, int *maxWidth) const;
+    void getMaxWidth(std::vector<QString> toCompare, qreal *maxWidth) const;
     void setFontItalic(bool enable, QPainter *painter);
     void mouseDoubleClickEvent(QGraphicsSceneMouseEvent *event);
     void mouseReleaseEvent(QGraphicsSceneMouseEvent *event);
